# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DayChallenge API",
        "description": "Habit Gamification & Personal Finance API",
        "contact": {
            "name": "DayChallenge Support",
            "email": "support@daychallenge.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "v1"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Local Server"
        }
    ],
    "paths": {
...
        "/finance/expenses": {
            "get": {
                "tags": [
                    "Finance"
                ],
                "summary": "List expenses",
                "description": "Retrieves a paginated list of expenses for the user and total spending in ETB.",
                "operationId": "0318f27df03c00ab7e51997e2de04833",
                "responses": {
                    "200": {
                        "description": "Expenses retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Expenses retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Expense"
                                            }
                                        },
                                        "spending": {
                                            "properties": {
                                                "total_spending": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 45000
                                                },
                                                "currency": {
                                                    "type": "string",
                                                    "example": "ETB"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Finance"
                ],
                "summary": "Create expense record",
                "description": "Creates a new expense record in ETB.",
                "operationId": "c86606c39dbbb8ed79c543c868ccb9ef",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreExpenseRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Expense created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Expense created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Expense"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failure",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/finance/expenses/{expense}": {
            "put": {
                "tags": [
                    "Finance"
                ],
                "summary": "Update expense record",
                "description": "Updates an existing expense record.",
                "operationId": "8f424d84d471180ff9ab86da6078293b",
                "parameters": [
                    {
                        "name": "expense",
                        "in": "path",
                        "description": "The ID of the expense",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreExpenseRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Expense updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Expense updated successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Expense"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized access",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Expense not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failure",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Finance"
                ],
                "summary": "Delete expense record",
                "description": "Deletes an expense record.",
                "operationId": "246ec1376a6d00e7f369813fba2957d4",
                "parameters": [
                    {
                        "name": "expense",
                        "in": "path",
                        "description": "The ID of the expense",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Expense deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Expense deleted successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized access",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Expense not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Expense": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "9b3d9d30-b8bd-474c-8822-67cc3b922a99"
                    },
                    "category": {
                        "type": "string",
                        "example": "Food & Dining"
                    },
                    "amount": {
                        "type": "number",
                        "format": "float",
                        "example": 4500
                    },
                    "expense_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-07-09"
                    },
                    "description": {
                        "type": "string",
                        "example": "Groceries from Shola Market",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-07-09T03:56:49.000000Z"
                    }
                },
                "type": "object"
            },
...
            "StoreExpenseRequest": {
                "required": [
                    "category_id",
                    "amount",
                    "expense_date"
                ],
                "properties": {
                    "category_id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "9b3d9d30-b8bd-474c-8822-67cc3b922a99"
                    },
                    "amount": {
                        "type": "number",
                        "format": "float",
                        "example": 1200,
                        "minimum": 0
                    },
                    "expense_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-07-09"
                    },
                    "description": {
                        "type": "string",
                        "example": "Lunch with clients"
                    }
                },
                "type": "object"
            },
...
            "ValidationErrorResponse": {
                "properties": {
                    "status": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "properties": {
                            "field_name": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                },
                                "example": [
                                    "The field_name field is required."
                                ]
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "UnauthorizedResponse": {
                "properties": {
                    "status": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            },
            "NotFoundResponse": {
                "properties": {
                    "status": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Resource not found."
                    }
                },
                "type": "object"
            },
            "ForbiddenResponse": {
                "properties": {
                    "status": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "This action is unauthorized."
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Bearer token in the format: Bearer <token>",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}